fix install.sh

+++ .dotfiles/install.sh
@@ -9,7 +9,7 @@ dotfiles () { git --git-dir=$GIT_DIR --work-tree=$HOME "$@"; }
 
 # backup
 for file in $(dotfiles ls-tree -r --name-only HEAD); do
-  mv -vf $file $file.bak 2>/dev/null
+  [ -e $file ] && mv -vf $file $file.bak 2>/dev/null
 done
 
 # checkout
